Move useFakeTranslations to testHelpers
Allow reuse outside of `pageflow` package.

REDMINE-17375
