+++ packages/pageflow/spec/support/index.js
@@ -3,6 +3,5 @@ export * from 'pageflow/testHelpers';
 export * from './fakeEventEmitter';
 export * from './setupGlobals';
 
-export * from './useFakeTranslations';
 export * from './useFakeXhr';
 export * from './useHtmlSandbox'
