+++ packages/pageflow/src/testHelpers/index.js
@@ -1,3 +1,4 @@
 export * from './dominos/editor';
 export * from './dominos/ui';
 export * from './factories';
+export * from './useFakeTranslations';
